add platform.js
