+++ src/Entities/Entity.js
@@ -8,34 +8,6 @@ export default class Entity extends Phaser.GameObjects.Sprite {
     this.scene.physics.world.enableBody(this, 0);
     this.setData('type', type);
     this.setData('isDead', false);
-  // }
 
-  // explode(canDestroy) {
-  //   if (!this.getData('isDead')) {
-  //     this.setTexture('explosion');
-  //     this.play('explosion');
-
-  //     // this.scene.sfx.explosions[Phaser.Math.Between(0,
-  //     //   this.scene.sfx.explosions.length - 1)].play();
-
-  //     if (this.shootTimer !== undefined) {
-  //       if (this.shootTimer) {
-  //         this.shootTimer.remove(false);
-  //       }
-  //     }
-
-  //     this.setAngle(0);
-  //     this.body.setVelocity(0, 0);
-
-  //     this.on('animationcomplete', () => {
-  //       if (canDestroy) {
-  //         this.destroy();
-  //       } else {
-  //         this.setVisible(false);
-  //       }
-  //     }, this);
-
-  //     this.setData('isDead', true);
-  //   }
-  // }
-}}
+  }
+}
